+++ motioneye/__init__.py
@@ -1,1 +1,1 @@
-VERSION = "0.43.1b32"
+VERSION = "0.43.1b33"